Reset wave 1 file to the original version
